download-to-gcs/requirements.txt
google-cloud-storage==2.5.0
requests==2.28.1